aws_test_cli/lambda/cdk_lambda_sam_docker/external_lib_lambda_sam/requirements.txt
aws-cdk-lib==2.85.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
